requirements.txt
asgiref==3.8.1
charset-normalizer==3.4.0
Django==5.1.4
django-cors-headers==4.6.0
django-dotenv==1.4.2
psycopg2==2.9.10
djangorestframework==3.15.1
djangorestframework_simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.2
sqlparse==0.5.3
typing_extensions==4.12.2
drf-spectacular==0.28.0
